'profile_view_vote_star_rank'

+++ home/models.py
@@ -10,7 +10,7 @@ class Profile(models.Model):
     address = models.CharField(max_length=100,blank=True)
     mobile = models.CharField(max_length=60,blank=True)
     facebook = models.URLField(max_length=60,blank=True)
-    total_points = models.DecimalField(max_digits=15, decimal_places=2,default=0)
+    total_points = models.IntegerField(default=0)
     user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,related_name="profile")
 
     def __str__(self):
